Refactor login screen layout for improved structure and user experience; introduce SelectAccountTypeScreen for account type selection with enhanced UI elements and logo integration.
